将所有的id由string改为long型

+++ src/main/java/net/jeeshop/biz/article/service/ArticleCatalogService.java
@@ -27,7 +27,7 @@ public class ArticleCatalogService implements BaseService<ArticleCatalog> {
 
 
     @Override
-    public int insert(ArticleCatalog articleCatalog) {
+    public long insert(ArticleCatalog articleCatalog) {
         return articleCatalogMapper.insert(articleCatalog);
     }
